Add features of v6 & add import command (#31)
* Begin

* Still some trouble with overwriting navigation

* Big cleanup of default

* Fix Navigation default

* Move allow_printing to shared navigation data

* Update CHANGELOG.md

* Update dockerfile to rust 1.53

* Fix warnings

* Add builtin_constants and custom_constants in question description

* Add skill tags for topics

* Begin of numbas-to-rumbas

Changes to numbas crate so parsing diagnosys works

* Fix compilation of rumbas crate

Changes to OptionalOverwrite:
  - Splitted to EmptyFields and OptionalOverwrite
  - Make OptionalOverwrite generic (instead of adjacent type)

* Begin of converting exams: all but questions

* Converting exams: topics & learning objectives + question_group basics

* Most fields of question are converted

* Fix converting functions + Begin converting parts

* Improve conversion of JME part

* Fix deserialization of FileStrings & InputStrings

* Fix parsing of diagnostic_topic_names (had an extra space in front)

* Extract conversion of common part fields

* JME part conversion working

* NumberEntry part conversion working

* PatternMatch part conversion working

* ChooseOne part conversion (almost totally) working

* ChooseMultiple part conversion (almost totally) working

* Fix typo

* Improve conversion of ChooseOne & ChooseMultiple: Also create ItemBased

* Information part conversion

* GapFill part conversion

* MatchItems part conversion

* Show where invalid things are (instead of ugly serde error) & write as ExamFileType

* Change EmptyFields to CheckRumbas

* Fix numbas-to-rumbas: no value in TranslatableString anymore

* Create seperate yaml for questions instead of inline (which doesn't work)

* Write yaml to files

* Basic question name sanitation

* Try improving workflows

* Ignore a numbas test

* Fix rumbas tests

* Try fixing todo check

* Fix dockerfile

* Run rust workflow on all pr's

* Add list_todos.sh & let ci also check todo's in deps

* Fix building docker image

* Also convert normal exams

* Fix some stuff so the getting started example can be converted

* Add DEFAULTS constant with numbas defaults

* Fix extract_checking_type

* Fixed everything needed for exam-14065 (geogebra, parts extension & matrix, custom_part_type, ...) except resources

* Add support for resources

* Fix some rumbas clippy warnings

* Fix all clippy warnings in rumbas crate

* Fix all clippy warnings in numbas & numbas-to-rumbas

* Cleanup of numbas-to-rumbas to make extraction possible

* Move numbas-to-rumbas to import subcommand of rumbas

* Cleanup workflows

* Fix clippy warnings

* Add rumbas init & cli-tests script

* Also run cli-tests on ci

* Split cli test for rumbas test on ci

* Fix cli-tests

* Fix tests (docker & cli-tests)

* Update README & GETTING STARTED

* Add use_field_init_shorthand back to rustfmt.toml in rumbas crate

* Fix JME answer_check field

* Remove some wrong todo's

* Fix custom_part_types (in questions instead of exam) & unique resources & custom_part_types

* Fix variables (remove Value's as root element in enum)

* Fix compile s & z options

* Remove old TODOs

* Document aliases

* Add reverse to CHANGELOG

* Cleanup TODO's

* Fix CHANGELOG

+++ rumbas/src/main.rs
@@ -1,19 +1,22 @@
-use crate::data::default::combine_with_default_files;
-use crate::data::to_numbas::ToNumbas;
+use rumbas::data::default::combine_with_default_files;
+use rumbas::data::to_numbas::ToNumbas;
 use std::env;
 use std::path::Path;
 #[macro_use]
 extern crate clap;
 use clap::{crate_version, App};
+mod import;
+mod init;
 
-mod data;
+use import::import;
+use init::init;
 
 /// The name of the local folder used as cache
 /// It caches the .exam files that are given to Numbas.
-const CACHE_FOLDER: &'static str = ".rumbas";
+const CACHE_FOLDER: &str = ".rumbas";
 
 /// The name of the local folder used for the output.
-const OUTPUT_FOLDER: &'static str = "_output";
+const OUTPUT_FOLDER: &str = "_output";
 
 /// The main cli function
 /// # Requirements
@@ -25,10 +28,20 @@ const OUTPUT_FOLDER: &'static str = "_output";
 /// The `_output` folder will contain the generated HTML files.
 /// Host these with a webserver.
 fn main() {
-    let numbas_path = env::var("NUMBAS_FOLDER").expect("NUMBAS_FOLDER to be set");
     let yaml = load_yaml!("cli.yml");
     let matches = App::from_yaml(yaml).version(crate_version!()).get_matches();
 
+    if let Some(matches) = matches.subcommand_matches("import") {
+        import(&matches)
+    } else if let Some(matches) = matches.subcommand_matches("compile") {
+        compile(&matches)
+    } else if let Some(matches) = matches.subcommand_matches("init") {
+        init(&matches)
+    }
+}
+fn compile(matches: &clap::ArgMatches) {
+    let numbas_path = env::var(rumbas::NUMBAS_FOLDER_ENV)
+        .expect(&format!("{} to be set", rumbas::NUMBAS_FOLDER_ENV)[..]);
     let path = Path::new(matches.value_of("EXAM_OR_QUESTION_PATH").unwrap());
     println!("{:?}", path.display());
     if path.is_absolute() {
@@ -55,14 +68,14 @@ fn main() {
         _ => 3,
     };*/
 
-    let exam_result = data::exam::Exam::from_file(path);
+    let exam_result = rumbas::data::exam::Exam::from_file(path);
     match exam_result {
         Ok(mut exam) => {
             combine_with_default_files(path, &mut exam);
-            if exam.locales.0.is_none() {
+            if exam.locales().0.is_none() {
                 eprintln!("Locales not set!");
             } else {
-                for locale_item in exam.locales.clone().unwrap().iter() {
+                for locale_item in exam.locales().unwrap().iter() {
                     let locale = locale_item.clone().unwrap().name.unwrap();
                     let numbas = exam.to_numbas(&locale);
                     match numbas {
@@ -80,29 +93,27 @@ fn main() {
                                 .canonicalize()
                                 .unwrap()
                                 .join(path.with_extension(output_extension));
-                            let numbas_output_path = if output_extension == "" {
+                            if output_extension.is_empty() {
                                 // Remove current folder
                                 std::fs::remove_dir_all(&absolute_path).unwrap_or(()); //If error, don't mind
                                                                                        // Create folder
                                 std::fs::create_dir_all(&absolute_path)
                                     .expect("Failed creating folder for output");
-                                absolute_path
                             } else {
                                 std::fs::create_dir_all(&absolute_path.parent().unwrap())
                                     .expect("Failed creating folder for output");
-                                absolute_path
                             };
+                            let numbas_output_path = absolute_path;
                             //println!("{}", numbas_output_path.display());
 
-                            res.write(&numbas_exam_path.to_str().unwrap());
+                            res.write(numbas_exam_path.to_str().unwrap());
 
-                            let numbas_settings = exam.numbas_settings.clone().unwrap();
+                            let numbas_settings = exam.numbas_settings().clone().unwrap();
 
                             let output = std::process::Command::new("python3")
                                 .current_dir(numbas_path.clone())
                                 .arg("bin/numbas.py")
                                 .arg("-l")
-                                //TODO.arg(&numbas_settings.locale.unwrap().to_str())
                                 .arg(locale_item.unwrap().numbas_locale.clone().unwrap().to_str())
                                 .arg("-t")
                                 .arg(numbas_settings.theme.unwrap())
@@ -112,19 +123,34 @@ fn main() {
                                 .arg(numbas_exam_path.canonicalize().unwrap()) //TODO?
                                 .output()
                                 .expect("failed to execute numbas process");
-                            if output.stdout.len() > 0 {
+                            if !output.stdout.is_empty() {
                                 println!("{}", std::str::from_utf8(&output.stdout).unwrap());
                             }
-                            if output.stderr.len() > 0 {
+                            if !output.stderr.is_empty() {
                                 eprintln!("{}", std::str::from_utf8(&output.stderr).unwrap());
+                                std::process::exit(1)
                             }
                         }
-                        Err(missing_fields) => {
+                        Err(check_result) => {
+                            let missing_fields = check_result.missing_fields();
+                            let invalid_fields = check_result.invalid_fields();
                             println!(
-                                "Missing fields ({}):\n{}",
+                                "Error when processing locale {}.\nFound {} missing fields:\n{}\nFound {} invalid fields:\n{}",
                                 locale,
-                                missing_fields.join("\n")
+                                missing_fields.len(),
+                                missing_fields
+                                    .iter()
+                                    .map(|f| f.to_string())
+                                    .collect::<Vec<_>>()
+                                    .join("\n"),
+                                invalid_fields.len(),
+                                invalid_fields
+                                    .iter()
+                                    .map(|f| f.to_string())
+                                    .collect::<Vec<_>>()
+                                    .join("\n")
                             );
+                            std::process::exit(1)
                         }
                     }
                 }
